service/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
flake8==6.0.0
flower==1.2.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
phonenumberslite==8.13.1
prometheus-client==0.15.0
prompt-toolkit==3.0.33
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
